requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
telethon==1.30.3
aiohttp==3.8.6
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
python-telegram-bot==20.6 